+++ nash-protocol/src/protocol/place_orders/request.rs
@@ -10,20 +10,8 @@ use super::types::{
 use tokio::sync::RwLock;
 use std::sync::Arc;
 
-use serde::{Serialize, Deserialize};
 use std::collections::HashMap;
-
-/// The form in which queries are sent over HTTP in most implementations. This will be built using the [`GraphQLQuery`] trait normally.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct MultiQueryBody {
-    /// The values for the variables. They must match those declared in the queries. This should be the `Variables` struct from the generated module corresponding to the query.
-    pub variables: HashMap<String, serde_json::Value>,
-    /// The GraphQL query, as a string.
-    pub query: String,
-    /// The GraphQL operation name, as a string.
-    #[serde(rename = "operationName")]
-    pub operation_name: &'static str,
-}
+use crate::protocol::multi_request::MultiQueryBody;
 
 impl LimitOrdersRequest {
     // Buy or sell `amount` of `A` in price of `B` for an A/B market. Returns a builder struct
